Audio classes moved to seperate files.
